frameworks/fm_autogpt/pyproject.toml
[project]
name = "fm-autogpt"
version = "0.1.0"
description = "AutoGPT framework implementation for BBH benchmarking"
requires-python = ">=3.10"
dependencies = [
    "AutoGPT-Forge>=0.1.4",
    "openai>=1.99.0",
    "datasets==3.6.0",
    "requests>=2.25.0",
    "pyyaml>=6.0",
]
setup = "ready"